float to double ; fixed jacobi

+++ src/integration/C/approx_e.c
@@ -3,7 +3,7 @@
 
 double approx_e(long n, short parallel) {
     double e = 0.0;
-    float factorial = 1.0;
+    double factorial = 1.0;
 
     if (parallel) {
         #pragma omp parallel for reduction(+:e)
