+++ src/linear_algebra/C/kronecker.c
@@ -16,23 +16,23 @@
 #define M 32
 #define N 32
 
-void print_matrix(int m, int n, float[m][n]);
-void gen_matrix(int m, int n, float A[m][n]);
+void print_matrix(int m, int n, double[m][n]);
+void gen_matrix(int m, int n, double A[m][n]);
 int rand_lim(int limit);
-void arr_alloc_2 (size_t m, size_t n, float(**aptr)[m][n]);
-void arr_alloc_4 (size_t a, size_t b, size_t c, size_t d, float(**aptr)[a][b][c][d]);
+void arr_alloc_2 (size_t m, size_t n, double(**aptr)[m][n]);
+void arr_alloc_4 (size_t a, size_t b, size_t c, size_t d, double(**aptr)[a][b][c][d]);
 
 int main(void) {
     printf("   M = %d, N = %d\n\n", M, N);
     double time;
 
-    float A[M][N];
+    double A[M][N];
     gen_matrix(M, N, A);
 
-    float B[M][N];
+    double B[M][N];
     gen_matrix(M, N, B);
 
-    float (result)[M][N][M][N];
+    double (result)[M][N][M][N];
 
     printf("   Kronecker Product (serial)\n");
     time = omp_get_wtime();
@@ -49,7 +49,7 @@ int main(void) {
     return 0;
 }
 
-void print_matrix(int m, int n, float A[m][n]) {
+void print_matrix(int m, int n, double A[m][n]) {
     for (int i = 0; i < m; i++) {
         for (int j = 0; j < n; j++) {
             printf("%.1f  ", A[i][j]);
@@ -59,10 +59,10 @@ void print_matrix(int m, int n, float A[m][n]) {
     printf("\n");
 }
 
-void gen_matrix(int m, int n, float A[m][n]) {
+void gen_matrix(int m, int n, double A[m][n]) {
     for (int i = 0; i < m; i++) {
         for (int j = 0; j < n; j++) {
-            A[i][j] = (float) rand_lim(LIMIT);
+            A[i][j] = (double) rand_lim(LIMIT);
         }
     }
 }
@@ -78,13 +78,13 @@ int rand_lim(int limit) {
     return retval;
 }
 
-void arr_alloc_2 (size_t m, size_t n, float(**aptr)[m][n])
+void arr_alloc_2 (size_t m, size_t n, double(**aptr)[m][n])
 {
-  *aptr = malloc(sizeof(float[m][n]) ); // allocate a true 2D array
+  *aptr = malloc(sizeof(double[m][n]) ); // allocate a true 2D array
   assert(*aptr != NULL);
 }
 
-void arr_alloc_4(size_t a, size_t b, size_t c, size_t d, float(**aptr)[a][b][c][d]) {
-    *aptr = malloc(sizeof(float[a][b][c][d]) ); // allocate a true 2D array
+void arr_alloc_4(size_t a, size_t b, size_t c, size_t d, double(**aptr)[a][b][c][d]) {
+    *aptr = malloc(sizeof(double[a][b][c][d]) ); // allocate a true 2D array
     assert(*aptr != NULL);
 }
