+++ src/linear_algebra/C/power_method.c
@@ -18,21 +18,21 @@
 #define N 256
 #define MAX_ITERS 1000
 
-void print_vector(int n, float[]);
-void gen_vector(int n, float v[]);
-void print_matrix(int m, int n, float[m][n]);
-void gen_matrix(int m, int n, float A[m][n]);
+void print_vector(int n, double[]);
+void gen_vector(int n, double v[]);
+void print_matrix(int m, int n, double[m][n]);
+void gen_matrix(int m, int n, double A[m][n]);
 int rand_lim(int limit);
-void arr_alloc (size_t m, size_t n, float(**aptr)[m][n]);
+void arr_alloc (size_t m, size_t n, double(**aptr)[m][n]);
 
 int main(void) {
     printf("   M = %d, N = %d\n\n", M, N);
     double time;
 
-    float eigen_value;
-    float eigen_vector[N];
+    double eigen_value;
+    double eigen_vector[N];
 
-    float (*A)[M][N];
+    double (*A)[M][N];
     arr_alloc(M, N, &A);
     gen_matrix(M, N, *A);
 
@@ -65,20 +65,20 @@ int main(void) {
     return 0;
 }
 
-void print_vector(int n, float v[]) {
+void print_vector(int n, double v[]) {
     for (int i = 0; i < n; i++) {
         printf("%f  ", v[i]);
     }
     printf("\n");
 }
 
-void gen_vector(int n, float v[]) {
+void gen_vector(int n, double v[]) {
     for (int i = 0; i < n; i++) {
-        v[i] = (float) rand_lim(LIMIT);
+        v[i] = (double) rand_lim(LIMIT);
     }
 }
 
-void print_matrix(int m, int n, float A[m][n]) {
+void print_matrix(int m, int n, double A[m][n]) {
     for (int i = 0; i < m; i++) {
         for (int j = 0; j < n; j++) {
             printf("%f  ", A[i][j]);
@@ -88,10 +88,10 @@ void print_matrix(int m, int n, float A[m][n]) {
     printf("\n");
 }
 
-void gen_matrix(int m, int n, float A[m][n]) {
+void gen_matrix(int m, int n, double A[m][n]) {
     for (int i = 0; i < m; i++) {
         for (int j = 0; j < n; j++) {
-            A[i][j] = (float) rand_lim(LIMIT);
+            A[i][j] = (double) rand_lim(LIMIT);
         }
     }
 }
@@ -107,8 +107,8 @@ int rand_lim(int limit) {
     return retval;
 }
 
-void arr_alloc (size_t m, size_t n, float(**aptr)[m][n])
+void arr_alloc (size_t m, size_t n, double(**aptr)[m][n])
 {
-  *aptr = malloc(sizeof(float[m][n]) ); // allocate a true 2D array
+  *aptr = malloc(sizeof(double[m][n]) ); // allocate a true 2D array
   assert(*aptr != NULL);
 }
